updated behaviors and added aoe and piercable

+++ src/gameEngine/actors/ProjectileInfo.java
@@ -12,11 +12,11 @@ public class ProjectileInfo {
     private IBehavior myMovement;
     private List<IBehavior> myOnHitEffects;
     private List<String> myEnemyTypes;
-    private int myDamage;
+    private double myDamage;
 
 
 
-    public ProjectileInfo(String image, int damage, Map<String,IBehavior> map, List<String> types){
+    public ProjectileInfo(String image, double damage, Map<String,IBehavior> map, List<String> types){
         myDamage=damage;
         myImage = image;
         myOnHitEffects=new ArrayList<>();
@@ -30,7 +30,7 @@ public class ProjectileInfo {
         myEnemyTypes=types;
 
     }
-    public ProjectileInfo(String image, int damage, IBehavior move, List<IBehavior> list, List<String> types){
+    public ProjectileInfo(String image, double damage, IBehavior move, List<IBehavior> list, List<String> types){
         myImage=image;
         myDamage=damage;
         myOnHitEffects=list;
@@ -54,11 +54,11 @@ public class ProjectileInfo {
         return (BaseMovementBehavior)myMovement;
     }
 
-    public int getMyDamage () {
+    public double getMyDamage () {
         return myDamage;
     }
 
-    public void setMyDamage (int myDamage) {
+    public void setMyDamage (double myDamage) {
         this.myDamage = myDamage;
     }
     public ProjectileInfo copy(){
